Project

+++ 21_HC-SR04_ULTRASONIC_SENSOR/HC_SR04.h
@@ -8,6 +8,6 @@
 
 void HCSR04_INIT(void);
 void Pulse(void);
-float Range(void);
+int Range(void);
 
 #endif // __HCSR04_H__
